Separate out caseload components
