+++ src/controllers/caseload/caseload.service.ts
@@ -150,7 +150,6 @@ export class CaseloadService {
         }),
       );
     }
-    console.log(getRequestSpecs);
     return getRequestSpecs;
   }
